Delete habits

+++ HabitTrackerProject/HabitTracker/ConsoleUI.cs
@@ -119,10 +119,20 @@ public class ConsoleUI
                     if (choice == "Go back") {
                         break;
                     }
-                    Habit habitChoice = new Habit(choice);
-                    dataManager.deleteHabit(habitChoice);
 
-                    AnsiConsole.WriteLine($"{habitChoice} successfully deleted.");
+                    string deleteChoice = AnsiConsole.Prompt(
+                        new SelectionPrompt<string>()
+                            .Title($"[red]Are you sure you want to delete {choice}?[/]")
+                            .AddChoices(new[] {
+                                "Yes",
+                                "No"
+                            }));
+                    if (deleteChoice == "Yes") {
+                        Habit habitChoice = new Habit(choice);
+                        dataManager.deleteHabit(habitChoice);
+
+                        AnsiConsole.WriteLine($"{habitChoice} successfully deleted.");
+                    }
                 }
             } else if (command == "View report") {
                 
@@ -132,28 +142,48 @@ public class ConsoleUI
                             .Title("There are no stored habits.")
                             .AddChoices(GetHabitChoices()));
                 } else {
-                    choice = AnsiConsole.Prompt(
-                    new SelectionPrompt<string>()
-                        .Title("Choose habit")
-                        .AddChoices(GetHabitChoices()));
-                }
-                Table table = new Table();
-                table.AddColumn("Habit");
-                table.AddColumn("Goal");
-                table.AddColumn("Done");
-                table.AddColumn("Status");
+                    Table table = new Table();
+                    table.AddColumn("Habit");
+                    table.AddColumn("Goal");
+                    table.AddColumn("Done");
+                    table.AddColumn("Status");
 
-                foreach (Habit habit in dataManager.Habits) {
-                    string status;
-                    
-                    if (habit.getDone() == 0) {
-                        status = "[red]Not Started[/]";
-                    } else {
-                        status = habit.isCompleted() ? "[green]Complete[/]" : "[red]Incomplete[/]";
+                    foreach (Habit habit in dataManager.Habits) {
+                        string status;
+                        
+                        if (habit.getDone() == 0) {
+                            status = "[red]Not Started[/]";
+                        } else {
+                            status = habit.isCompleted() ? "[green]Complete[/]" : "[red]Incomplete[/]";
+                        }
+                        table.AddRow(habit.Name, Convert.ToString(habit.getGoal()), Convert.ToString(habit.getDone()), status);
                     }
-                    table.AddRow(habit.Name, Convert.ToString(habit.getGoal()), Convert.ToString(habit.getDone()), status);
+                    AnsiConsole.Write(table);
                 }
-                AnsiConsole.Write(table);
+                
+            } else if (command == "Delete all habits") {
+                    if (dataManager.Habits.Count == 0) {
+                        choice = AnsiConsole.Prompt(
+                            new SelectionPrompt<string>()
+                                .Title("There are no stored habits.")
+                                .AddChoices(GetHabitChoices()));
+                    } else {
+                        string deleteChoice = AnsiConsole.Prompt(
+                        new SelectionPrompt<string>()
+                            .Title("[red]Are you sure you want to delete all habits?[/]")
+                            .AddChoices(new[] {
+                                "Yes",
+                                "No"
+                            }));
+                        if (deleteChoice == "Yes") {
+                            dataManager.DeleteAllHabits();
+                            AnsiConsole.WriteLine("All habits have been deleted.");
+                        }
+                        
+                    }
+                
+
+                
             }
             
             command = ShowMainMenu();
@@ -170,6 +200,7 @@ public class ConsoleUI
                     "Update habit goals",
                     "Input habit",
                     "Delete habit",
+                    "Delete all habits",
                     "View report",
                     "Exit program"
                 }));
